+++ semcore/select/src/Select.jsx
@@ -53,10 +53,14 @@ class RootSelect extends Component {
     i18n: localizedMessages,
     locale: 'en',
   });
+  state = {
+    focusCatch: false,
+  };
 
   firstSelectedOptionRef = React.createRef();
 
   triggerRef = React.createRef();
+  focusableTriggerReturnFocusToRef = React.createRef();
 
   isScrolledToFirstOption = false;
 
@@ -104,6 +108,7 @@ class RootSelect extends Component {
       children: this.renderChildrenTrigger(value, options),
       getI18nText,
       ref: this.triggerRef,
+      focusableTriggerReturnFocusToRef: this.focusableTriggerReturnFocusToRef,
     };
   }
 
@@ -117,6 +122,11 @@ class RootSelect extends Component {
       'aria-label': 'List of options',
     };
   }
+  getPopperProps() {
+    return {
+      focusableTriggerReturnFocusToRef: this.focusableTriggerReturnFocusToRef,
+    };
+  }
 
   getMenuProps() {
     const { uid, getI18nText, multiselect } = this.asProps;
@@ -126,11 +136,12 @@ class RootSelect extends Component {
       id: `igc-${uid}-list`,
       role: 'listbox',
       'aria-label': getI18nText('optionsList'),
+      focusableTriggerReturnFocusToRef: this.focusableTriggerReturnFocusToRef,
     };
   }
 
   getOptionProps(props) {
-    const { value, uid } = this.asProps;
+    const { value } = this.asProps;
     const selected = isSelectedOption(value, props.value);
     const other = {};
     this._optionSelected = selected;
@@ -142,7 +153,6 @@ class RootSelect extends Component {
     return {
       selected,
       'aria-selected': selected ? 'true' : 'false',
-      id: `igc-${uid}-option-${props.value}`,
       role: 'option',
       onClick: this.bindHandlerOptionClick(props.value),
       ...other,
@@ -202,7 +212,7 @@ class RootSelect extends Component {
 
   bindHandlerOptionClick = (optionValue) => (e) => {
     let newValue = optionValue;
-    const { value, multiselect } = this.asProps;
+    const { value, multiselect, interaction } = this.asProps;
     if (Array.isArray(value)) {
       if (value.includes(optionValue)) {
         newValue = value.filter((v) => v !== optionValue);
@@ -214,7 +224,15 @@ class RootSelect extends Component {
     if (!multiselect) {
       this.handlers.visible(false);
       setTimeout(() => {
-        // this.triggerRef.current?.focus();
+        if (interaction !== 'focus') {
+          this.triggerRef.current?.focus();
+          return;
+        }
+        if (this.focusableTriggerReturnFocusToRef.current) {
+          this.focusableTriggerReturnFocusToRef.current.focus();
+          return;
+        }
+        this.setState({ focusCatch: true });
       }, 0);
     }
   };
@@ -243,6 +261,13 @@ class RootSelect extends Component {
       return (
         <Root render={DropdownMenu}>
           <Select.Trigger {...other} role="combobox" />
+          {this.state.focusCatch && (
+            <div
+              tabIndex={0}
+              ref={(node) => node?.focus()}
+              onBlur={() => this.setState({ focusCatch: false })}
+            />
+          )}
           <Select.Menu>
             {options.map((option) => {
               return (
@@ -265,6 +290,13 @@ class RootSelect extends Component {
     return (
       <Root render={DropdownMenu}>
         <Children />
+        {this.state.focusCatch && (
+          <div
+            tabIndex={0}
+            ref={(node) => node?.focus()}
+            onBlur={() => this.setState({ focusCatch: false })}
+          />
+        )}
       </Root>
     );
   }
@@ -287,6 +319,7 @@ function Trigger({
   $hiddenRef,
   tag: Tag = ButtonTrigger,
   getI18nText,
+  focusableTriggerReturnFocusToRef,
 }) {
   const hasInputTrigger = isInputTriggerTag(Tag);
   const SSelectTrigger = Root;
@@ -298,6 +331,7 @@ function Trigger({
       placeholder={getI18nText('selectPlaceholder')}
       aria-autocomplete={(hasInputTrigger && 'list') || undefined}
       role={(hasInputTrigger && 'combobox') || undefined}
+      focusableTriggerReturnFocusToRef={focusableTriggerReturnFocusToRef}
     >
       {addonTextChildren(
         Children,