+++ semcore/dropdown-menu/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.7.0] - 2023-05-10
+
+### Fixed
+
+- Improved `<Popper />` integration.
+- Fixed `aria-activedescendant` value.
+
 ## [3.7.2] - 2023-05-25
 
 ### Changed