+++ semcore/popper/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.18.0] - 2023-06-10
+
+- Fixed corner case of focus return with `interaction=focus`.
+- Added `disableEnforceFocus` api prop.
+
 ## [4.17.15] - 2023-05-29
 
 ### Changed
@@ -39,6 +44,8 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - Version patch update due to children dependencies update (`@semcore/animation` [1.10.8 ~> 1.10.9], `@semcore/flex-box` [4.7.22 ~> 4.7.23], `@semcore/neighbor-location` [3.1.28 ~> 3.1.29], `@semcore/outside-click` [2.5.27 ~> 2.5.28], `@semcore/utils` [3.50.6 ~> 3.50.7]).
 
 ## [4.17.9] - 2023-05-11
+
+## [4.17.8] - 2023-05-04
 
 ### Changed