+++ semcore/select/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.7.0] - 2023-05-10
+
+### Fixed
+
+- Fixed browser focus cursor position after selecting value in non multiple choice `<Select />`.
+
 ## [3.6.16] - 2023-05-25
 
 ### Changed